+++ app/views/web/admin/jobs/_job.html.haml
